pyproject.toml
[project]
name = "RemarkableOCR"
version = "2024.09.0"
authors = [
    { name="Starlight", email="david@starlight.ai" },
]
maintainers = [
    { name="Starlight", email="david@starlight.ai" },
]
description = "RemarkableOCR is a simple ocr tool with improved data, analytics, and rendering tools."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
keywords = []

dependencies = [
    "pytesseract",
    "Pillow",
    "numpy",
    "scikit-image"
]


[project.urls]
Homepage = "https://github.com/markelwin/RemarkableOCR"
Repository = "https://github.com/markelwin/RemarkableOCR.git"
Issues = "https://github.com/markelwin/RemarkableOCR/issues"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["remarkable"]

[tool.pytest.ini_options]
testpaths = ["tests"]